+++ crates/bonsaidb-server/src/cli/mod.rs
@@ -26,7 +26,7 @@ pub enum Command<B: Backend = NoBackend> {
 impl<B: Backend> Command<B> {
     /// Executes the command.
     pub async fn execute(
-        &self,
+        self,
         configuration: ServerConfiguration<B>,
     ) -> Result<(), BackendError<B::Error>> {
         let server = CustomServer::<B>::open(configuration).await?;
@@ -34,7 +34,7 @@ impl<B: Backend> Command<B> {
     }
 
     /// Executes the command on `server`.
-    pub async fn execute_on(&self, server: CustomServer<B>) -> Result<(), BackendError<B::Error>> {
+    pub async fn execute_on(self, server: CustomServer<B>) -> Result<(), BackendError<B::Error>> {
         match self {
             Self::Certificate(command) => command.execute(&server).await,
             Self::Serve(command) => command.execute(&server).await,